% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoline_shiny}
\alias{autoline_shiny}
\title{Automatically wrap ggplot2 texts depending on rendered size}
\usage{
autoline_shiny(session, text, plot_name, font_size = 14, n = 1.6)
}
\arguments{
\item{session}{Shiny's session object.}

\item{text}{Character. Text for the title or subtitle or any text
to be wrapped.}

\item{plot_name}{Character. The name of the output plot.}

\item{font_size, n}{Numeric. Values to help calculate a proxy of length.}
}
\value{
Character. String with wrapped values.
}
\description{
This function automatically adapts the length of a title or subtitle
text so that it wraps long texts in separate lines. This is useful
to avoid trimming long texts when the rendered shiny size is shorter.
}
