% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{custom_brand}
\alias{custom_brand}
\title{Brand Customize for Shiny's Header}
\usage{
custom_brand(title = "MyLareShiny",
  site = "https://github.com/laresbernardo/lareshiny", favicon = NULL,
  font = "Montserrat", logosrc = NULL, loadingsrc = NULL,
  height = NULL, width = NULL, text = Sys.Date())
}
\arguments{
\item{title}{Character. Your Shinys title (displayed in Navigators' tab)}

\item{site}{Character. URL for your site (opens when logo is clicked)}

\item{favicon}{Character. Image for your favicon. Save file in www directory}

\item{font}{Character. Font for whole Shiny app. Use Google Fonts names}

\item{logosrc, loadingsrc}{Character. Logo image and loading image}

\item{height, width}{Integer. Logo image dimentions}

\item{text}{Character. Text displayed in top right corner}
}
\description{
This function personalizes your Shiny dashboard's header with logo, 
links, favicon, font, and texts.
}
