% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{js_op}
\alias{js_op}
\title{Totals Row for datatables}
\usage{
js_op(column, operation, txt = "", signif = 3)
}
\arguments{
\item{column}{Integer. Starting from 0, which column to operate}

\item{operation}{Character. Select from sum, mean, count, custom}

\item{txt}{Character. Insert text before (or instead) operation}

\item{signif}{Integer. How many decimals to consider when operating}
}
\description{
Totals Row for datatables
}
