% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{js_op_aux}
\alias{js_op_aux}
\title{Helper for javascript code}
\usage{
js_op_aux(type, df = NULL)
}
\arguments{
\item{type}{Character. Select start, end or sketch}

\item{df}{data.frame. Needed for sketch only}
}
\description{
Helper for javascript code
}
