% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lareshiny.R
\docType{package}
\name{lareshiny}
\alias{lareshiny-package}
\alias{lareshiny}
\title{Lares Shiny Modules}
\description{
R library for quick useful Shiny modules
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laresbernardo/lareshiny}
  \item \url{https://laresbernardo.github.io/lareshiny/}
  \item Report bugs at \url{https://github.com/laresbernardo/lareshiny/issues}
}

}
\author{
Bernardo Lares (lareshiny@bervos.org)
}
