% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_exporter}
\alias{model_exporter}
\title{Export Results from h2o_automl()}
\usage{
model_exporter(model)
}
\arguments{
\item{model}{List. Result from h2o_automl()}
}
\description{
Friendly app that helps you export lares' h2o_automl() results
}
\examples{
if (interactive()) {
  model_exporter(model)
}
}
