% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{try_require}
\alias{try_require}
\title{Check if Specific Package is Installed}
\usage{
try_require(package, stop = TRUE)
}
\arguments{
\item{package}{Character. Name of the library}

\item{stop}{Boolean. Stop if not installed}
}
\description{
This function checks library dependencies
}
