% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{custom_footer}
\alias{custom_footer}
\title{Brand Customize for Shiny's Header}
\usage{
custom_footer(company = "MyCoolCompany",
  site = "https://github.com/laresbernardo/lareshiny",
  helpmail = "myemail@mydomain.com", type = 2)
}
\arguments{
\item{company, site}{Character. Your company's name and URL}

\item{helpmail}{Character. Developers mail}

\item{type}{Integer. Which footer type to use? Needs further developments}
}
\description{
This function personalizes your Shiny dashboard's header with logo, 
links, favicon, font, and texts.
}
