% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{custom_footer}
\alias{custom_footer}
\title{Custom Footer: Customize for Shiny's Footer}
\usage{
custom_footer(
  company = "MyCoolCompany",
  text = "ShinyDashboard",
  site = "https://github.com/laresbernardo/lareshiny",
  helpmail = "myemail@mydomain.com",
  type = 2
)
}
\arguments{
\item{company, text, site}{Character. Your company's name and URL and
the project's name to show in footer.}

\item{helpmail}{Character. Developers mail}

\item{type}{Integer. Which footer type to use? Needs further developments.}
}
\value{
A tags$footer formatted object
}
\description{
This function personalizes your Shiny dashboard's footer with your
company's name, link, help mail.
}
