% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixds.R
\name{matrixwd}
\alias{matrixwd}
\title{Shiny Path for Dev/Production in MatrixDS}
\usage{
matrixwd(shiny_project = "Shiny")
}
\arguments{
\item{shiny_project}{Character. Directory's name}
}
\description{
Shiny Path for Dev/Production in MatrixDS
}
