% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{module_login}
\alias{module_login}
\title{Login Module for Shiny}
\usage{
module_login(
  input,
  session,
  users = c("123", "321"),
  pwds = c("123", "321"),
  logo = NA,
  lang = "es",
  style = list(botton_txt_colour = "#FFFFFF", botton_bgd_colour = "#EBB600"),
  logged = FALSE,
  personal = "MacBook-Pro-de-Bernardo.local"
)
}
\arguments{
\item{input, session}{Internal}

\item{users, pwds}{Character Vector. User and password combinations accepted}

\item{logo}{Character. Select image for logo display. Host local
file in your www directory preferably}

\item{lang}{Character. Language. Currently accepted: es, en}

\item{style}{List. Possible values for styling the module such as
botton_txt_colour and botton_bgd_colour}

\item{logged}{Boolean. You might want to set to TRUE when developing
or testing your app so this module doesn't show up every time.
Check the personal parameter as well}

\item{personal}{Character. If you wish to auto-login when you are
in your personal instance or computer, set your Sys.info's nodename.}
}
\value{
A reactiveValues object
}
\description{
Login Module for Shiny with local User and Password. This must be
added in the server section (no need to add UI)
}
