% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{try_require}
\alias{try_require}
\title{Check if Specific Package is Installed}
\usage{
try_require(package, stop = TRUE)
}
\arguments{
\item{package}{Character. Name of the library}

\item{stop}{Boolean. Stop if not installed. If \code{FALSE} and 
library is not available, warning will be shown.}
}
\value{
No return value, called for side effects.
}
\description{
This function checks library dependencies
}
\examples{
# Check if library base is installed. If not, stop and show error
try_require("base", stop = TRUE)
# Check if library xxx is installed. If not, show warning
try_require("xxx", stop = FALSE)
}
